Extract common parts from full and light nodes
